requirements.txt
streamlit
pandas
pandas-ta
requests
yfinance
google-genai
python-dotenv
plotly